Move existing project components to the 'pending' directory in preparation for the active project page
