+++ src/js/client/components/projects/pending/projectCheckpoint.js
