+++ src/js/client/components/projects/pending/projectType.js
